persimon/persimmon-api-develop/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A sample fastapi project created with create-fastapi-project"
authors = [ "Your Name <your@email.com>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "app"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.1"
ruff = "^0.5.7"
black = "^24.8.0"

[tool.black]
line-length = 88
target-version = [ "py37", "py38", "py39", "py310", "py311",]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 88
exclude = [ ".git", "__pycache__", ".mypy_cache", ".pytest_cache",]
select = [ "E", "W", "F", "C", "B",]
ignore = [ "B904", "B006", "E501", "B008", "C901",]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [ "alembic", "__pycache__",]

[tool.build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
xgboost = "^2.1.1"
fastapi = {extras = ["all"], version = "^0.112.0"}
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.26"}
asyncer = "^0.0.7"
httpx = "^0.27.0"
pdfplumber = "^0.11.4"
sqlmodel = "^0.0.21"
asyncpg = "^0.29.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
psycopg2 = "^2.9.9"
nltk = "3.9.1"
pandas = "^2.2.2"
pypdf2 = "^3.0.1"
firebase-admin = "^6.5.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
tensorflow = "2.17.0"
tensorflow-io-gcs-filesystem = "0.31.0"
spacy = "^3.0.0"
google-generativeai = "^0.8.3"
validators = "^0.34.0"
tldextract = "^5.1.2"
reportlab = "^4.2.5"
python-docx = "^1.1.2"
aiofiles = "^24.1.0"
langchain-google-genai = "^2.0.1"
langchain = "^0.3.4"
tika = "^2.6.0"
boto3 = "^1.35.54"
beautifulsoup4 = "^4.12.3"
google-cloud-pubsub = "^2.27.1"
scikit-learn = "1.5.1"
tensorflow-intel = "^2.13.0"
fpdf = "^1.7.2"

[tool.ruff.per-file-ignores]
"__init__.py" = [ "F401",]
